AI Chat: organize types
