+++ apps/src/aichat/types/customizations.ts
@@ -1,0 +1,44 @@
+import {ValueOf} from '@cdo/apps/types/utils';
+import {AiChatModelIds} from '@cdo/generated-scripts/sharedConstants';
+
+/** Model customizations and model card information for aichat levels.
+ *  selectedModelId is a foreign key to ModelDescription.id */
+export interface AiCustomizations {
+  selectedModelId: ValueOf<typeof AiChatModelIds>;
+  temperature: number;
+  systemPrompt: string;
+  retrievalContexts: string[];
+  modelCardInfo: ModelCardInfo;
+}
+
+// Model customizations sent to backend for aichat levels - excludes modelCardInfo.
+// The customizations will be included in request to LLM endpoint.
+export type AichatModelCustomizations = Omit<AiCustomizations, 'modelCardInfo'>;
+
+export type FieldVisibilities = {[key in keyof AiCustomizations]: Visibility};
+
+/** Chat bot Model Card information */
+export interface ModelCardInfo {
+  botName: string;
+  description: string;
+  intendedUse: string;
+  limitationsAndWarnings: string;
+  testingAndEvaluation: string;
+  exampleTopics: string[];
+  isPublished: boolean;
+}
+
+/** Metadata about a given model, common across all aichat levels */
+export interface ModelDescription {
+  id: ValueOf<typeof AiChatModelIds>;
+  name: string;
+  overview: string;
+  trainingData: string;
+}
+
+// Visibility for AI customization fields set by levelbuilders.
+export enum Visibility {
+  HIDDEN = 'hidden',
+  READONLY = 'readonly',
+  EDITABLE = 'editable',
+}
